Add -l option and -o output file

+++ src/number.rs
@@ -13,7 +13,7 @@ pub fn parse_number_radix(
 ) -> Result<Token, Error> {
     let i =
         unsafe { str::from_utf8_unchecked(&input[*c_src - token_len as usize + 2..*c_src - 1]) };
-    let i = u32::from_str_radix(i, u32::from(base))?;
+    let i = u64::from_str_radix(i, u32::from(base))?;
     *c_src -= 1;
     Ok(Token::NumericLiteral(Number::new(i, 0, 1, base)))
 }
@@ -21,7 +21,7 @@ pub fn parse_number_radix(
 #[inline]
 pub fn parse_number(input: &[u8], c_src: &mut usize, token_len: u64) -> Result<Token, Error> {
     let i = unsafe { str::from_utf8_unchecked(&input[*c_src - token_len as usize..*c_src - 1]) };
-    let i = u32::from_str_radix(i, 10)?;
+    let i = u64::from_str_radix(i, 10)?;
     *c_src -= 1;
     Ok(Token::NumericLiteral(Number::new(i, 0, 1, 10)))
 }
@@ -45,10 +45,10 @@ pub fn parse_number_decimal(
         }
     }
     let integer = unsafe { str::from_utf8_unchecked(&input[*c_src - token_len as usize..i_point]) };
-    let integer = u32::from_str_radix(integer, 10)?;
+    let integer = u64::from_str_radix(integer, 10)?;
 
     let decimal = unsafe { str::from_utf8_unchecked(&input[i_point + 1..*c_src - 1]) };
-    let decimal = u32::from_str_radix(decimal, 10)?;
+    let decimal = u64::from_str_radix(decimal, 10)?;
 
     *c_src -= 1;
     Ok(Token::NumericLiteral(Number::new(integer, decimal, 1, 10)))
@@ -77,12 +77,12 @@ pub fn parse_exponent(input: &[u8], c_src: &mut usize, token_len: u64) -> Result
         let integer = unsafe {
             str::from_utf8_unchecked(&input[*c_src - token_len as usize..i_point.unwrap()])
         };
-        let integer = u32::from_str_radix(integer, 10)?;
+        let integer = u64::from_str_radix(integer, 10)?;
         let decimal = unsafe { str::from_utf8_unchecked(&input[i_point.unwrap() + 1..i_e]) };
-        (integer, u32::from_str_radix(decimal, 10)?)
+        (integer, u64::from_str_radix(decimal, 10)?)
     } else {
         let integer = unsafe { str::from_utf8_unchecked(&input[*c_src - token_len as usize..i_e]) };
-        let integer = u32::from_str_radix(integer, 10)?;
+        let integer = u64::from_str_radix(integer, 10)?;
         (integer, 0)
     };
